A complete overhaul.
The complete code is rewritten to match c++ programming standards. One of the major issues tackled was the unnecessary copies of the matrices. Now, most of the function returns just move the data from one object to another instead of copying it. This was achieved by move semantics available in c++11 standard.